move the certPool loader function into pkg/certs
